chore: split formatting plugin code
